Add ReturnUrl case class.
