+++ app/com/gu/identity/frontend/controllers/Application.scala
@@ -1,9 +1,8 @@
 package com.gu.identity.frontend.controllers
 
-import javax.inject.Inject
-
 import com.gu.identity.frontend.configuration.Configuration
 import com.gu.identity.frontend.logging.Logging
+import com.gu.identity.frontend.models.ReturnUrl
 import com.gu.identity.frontend.views.ViewRenderer.renderSignIn
 import play.api.i18n.{MessagesApi, I18nSupport}
 import play.api.mvc._
